Demo_copy
Mon Jan 22, 2024, as I am working on the login database I fixed the issue of the two login buttons. However, when trying to login with admin@gmail.com there is an issue. I have to comment out two sections of code from script.js in public and templates but as I do so and save.. when I deploy with firebase the updates are not showing in inspect and the website is showing the same error each time.